requirements.txt
# PDF text extraction for AI resume parsing
PyPDF2>=3.0.0
pdfminer.six>=20240706
pdfplumber>=0.11.4
Faker>=19.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.12.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
bcrypt==4.0.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.0.0
fastapi==0.103.1
frozenlist==1.7.0
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.5.0
passlib==1.7.4
Pillow==10.0.0
propcache==0.3.2
psycopg2-binary==2.9.7
pyasn1==0.6.1
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyJWT==2.10.1
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.6
redis==6.2.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.27.0
twilio==9.6.3
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.23.2
yarl==1.20.1
